+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
   * GET,PUT,DELETE `/1.1/groups/<group_uuid>`
   * PUT,DELETE `/1.1/groups/<group_uuid>/extensions/<extension_id>`
+  * PUT,DELETE `/1.1/groups/<group_uuid>/fallbacks`
 
 * The following resources have been deprecated:
 
   * GET,PUT,DELETE `/1.1/groups/<group_id>`
   * PUT,DELETE `/1.1/groups/<group_id>/extensions/<extension_id>`
+  * PUT,DELETE `/1.1/groups/<group_id>/fallbacks`
 
 
 # 21.03